requirements-test.txt
-e .[test]

pytest-xdist
pytest-cov
pytest-env
